Move configuration provider to its own module
