+++ src/configuration/configuration_provider.rs
@@ -1,0 +1,521 @@
+use crate::files::{FileManager, FileType};
+use serde::{Deserialize, Serialize};
+
+pub const WINDOW_MARGIN: &str = "window_margin";
+pub const ALLOW_SELECTING_SAME_CENTER_WINDOWS: &str = "allow_selecting_same_center_windows";
+pub const FORCE_USING_ADMIN_PRIVILEGES: &str = "force_using_admin_privileges";
+pub const ADDITIONAL_WORKSPACE_COUNT: &str = "additional_workspace_count";
+pub const ENABLE_FEATURES_USING_MOUSE: &str = "enable_features_using_mouse";
+pub const DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED: &str = "delay_in_ms_before_dragging_is_allowed";
+pub const ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE: &str = "allow_moving_cursor_after_open_close_or_minimise";
+pub const SCROLLING_ANIMATION_DURATION_IN_MS: &str = "animation_duration_in_ms";
+pub const SCROLLING_RECONCILIATION_INTERVAL_IN_MS: &str = "reconciliation_interval_in_ms";
+
+pub(super) const CONFIGURATION_FILE_NAME: &str = "randolf.toml";
+const CONFIGURATION_FILE_PREFIX: &str = "# This file is automatically generated and can be updated by you and by Randolf.\n\
+  # It stores your Randolf configuration. See https://github.com/kimgoetzke/randolf for details.\n\n";
+pub(super) const DEFAULT_WINDOW_MARGIN_VALUE: i32 = 20;
+const DEFAULT_DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED: i32 = 750;
+pub(super) const DEFAULT_SCROLLING_ANIMATION_DURATION_IN_MS: i32 = 120;
+pub(super) const DEFAULT_SCROLLING_RECONCILIATION_INTERVAL_IN_MS: i32 = 250;
+
+#[derive(Debug, Serialize, Deserialize, Default)]
+pub(super) struct Configuration {
+  pub(super) general: GeneralConfiguration,
+  #[serde(default)]
+  pub(super) layout: LayoutConfiguration,
+  #[serde(default)]
+  pub(super) spatial_layout: SpatialLayoutConfiguration,
+  #[serde(default)]
+  pub(super) scrolling_layout: ScrollingLayoutConfiguration,
+  #[serde(default)]
+  pub(super) hotkey: Vec<CustomHotkey>,
+  #[serde(default)]
+  pub(super) exclusion_settings: ExclusionSettings,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub(super) struct GeneralConfiguration {
+  #[serde(default = "default_window_margin")]
+  pub(super) window_margin: i32,
+  #[serde(default = "default_force_using_admin_privileges")]
+  pub(super) force_using_admin_privileges: bool,
+  #[serde(default = "default_additional_workspace_count")]
+  pub(super) additional_workspace_count: i32,
+  #[serde(default = "default_enable_features_using_mouse")]
+  pub(super) enable_features_using_mouse: bool,
+  #[serde(default = "default_delay_in_ms_before_dragging_is_allowed")]
+  pub(super) delay_in_ms_before_dragging_is_allowed: i32,
+  #[serde(default = "default_allow_moving_cursor_after_close_or_minimise")]
+  pub(super) allow_moving_cursor_after_open_close_or_minimise: bool,
+}
+
+pub(super) fn default_window_margin() -> i32 {
+  DEFAULT_WINDOW_MARGIN_VALUE
+}
+
+fn validate_window_margin(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  if !config_str.contains(WINDOW_MARGIN) {
+    warn!(
+      "[{}] was missing; adding it now with default value: {}",
+      WINDOW_MARGIN, DEFAULT_WINDOW_MARGIN_VALUE
+    );
+    configuration_provider.set_i32(WINDOW_MARGIN, DEFAULT_WINDOW_MARGIN_VALUE);
+  } else if configuration_provider.config.general.window_margin < 0 {
+    warn!(
+      "[{}] is negative, setting to default value: {}",
+      WINDOW_MARGIN, DEFAULT_WINDOW_MARGIN_VALUE
+    );
+    configuration_provider.set_i32(WINDOW_MARGIN, DEFAULT_WINDOW_MARGIN_VALUE);
+  }
+}
+
+fn default_force_using_admin_privileges() -> bool {
+  false
+}
+
+fn validate_force_using_admin_privileges(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  if !config_str.contains(FORCE_USING_ADMIN_PRIVILEGES) {
+    warn!(
+      "[{}] was missing; adding it now with default value: {}",
+      FORCE_USING_ADMIN_PRIVILEGES,
+      default_force_using_admin_privileges()
+    );
+    configuration_provider.set_bool(FORCE_USING_ADMIN_PRIVILEGES, default_force_using_admin_privileges());
+  }
+}
+
+pub(super) fn default_additional_workspace_count() -> i32 {
+  2
+}
+
+fn validate_workspace_count(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  if !config_str.contains(ADDITIONAL_WORKSPACE_COUNT) {
+    warn!(
+      "[{}] was missing; adding it now with default value: [{}]",
+      ADDITIONAL_WORKSPACE_COUNT,
+      default_additional_workspace_count()
+    );
+    configuration_provider.set_i32(ADDITIONAL_WORKSPACE_COUNT, default_additional_workspace_count());
+  } else if configuration_provider.config.general.additional_workspace_count < 0 {
+    warn!(
+      "[{}] is negative, setting to default value: [{}]",
+      ADDITIONAL_WORKSPACE_COUNT,
+      default_additional_workspace_count()
+    );
+    configuration_provider.set_i32(ADDITIONAL_WORKSPACE_COUNT, default_additional_workspace_count());
+  } else if configuration_provider.config.general.additional_workspace_count > 8 {
+    warn!(
+      "[{}] is larger than 8 which is not permitted, setting to 8",
+      ADDITIONAL_WORKSPACE_COUNT,
+    );
+    configuration_provider.set_i32(ADDITIONAL_WORKSPACE_COUNT, 8);
+  }
+}
+
+fn default_enable_features_using_mouse() -> bool {
+  true
+}
+
+fn validate_features_using_mouse(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  if !config_str.contains(ENABLE_FEATURES_USING_MOUSE) {
+    warn!(
+      "[{}] was missing; adding it now with default value: {}",
+      ENABLE_FEATURES_USING_MOUSE,
+      default_enable_features_using_mouse()
+    );
+    configuration_provider.set_bool(ENABLE_FEATURES_USING_MOUSE, default_enable_features_using_mouse());
+  }
+}
+
+fn default_delay_in_ms_before_dragging_is_allowed() -> i32 {
+  DEFAULT_DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED
+}
+
+fn validate_delay_in_ms_before_dragging_is_allowed(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  if !config_str.contains(DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED) {
+    warn!(
+      "[{}] was missing; adding it now with default value: {}",
+      DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED,
+      default_delay_in_ms_before_dragging_is_allowed()
+    );
+    configuration_provider.set_i32(
+      DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED,
+      default_delay_in_ms_before_dragging_is_allowed(),
+    );
+  }
+}
+
+fn default_allow_moving_cursor_after_close_or_minimise() -> bool {
+  true
+}
+
+fn validate_allow_moving_cursor_after_close_or_minimise(
+  config_str: &str,
+  configuration_provider: &mut ConfigurationProvider,
+) {
+  if !config_str.contains(ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE) {
+    warn!(
+      "[{}] was missing; adding it now with default value: {}",
+      ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE,
+      default_allow_moving_cursor_after_close_or_minimise()
+    );
+    configuration_provider.set_bool(
+      ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE,
+      default_allow_moving_cursor_after_close_or_minimise(),
+    );
+  }
+}
+
+impl Default for GeneralConfiguration {
+  fn default() -> Self {
+    Self {
+      window_margin: default_window_margin(),
+      force_using_admin_privileges: default_force_using_admin_privileges(),
+      additional_workspace_count: default_additional_workspace_count(),
+      enable_features_using_mouse: default_enable_features_using_mouse(),
+      delay_in_ms_before_dragging_is_allowed: default_delay_in_ms_before_dragging_is_allowed(),
+      allow_moving_cursor_after_open_close_or_minimise: default_allow_moving_cursor_after_close_or_minimise(),
+    }
+  }
+}
+
+#[derive(Debug, Serialize, Deserialize, Default)]
+pub(super) struct LayoutConfiguration {
+  #[serde(default)]
+  pub(super) default: Layout,
+  #[serde(default, skip_serializing_if = "Vec::is_empty")]
+  pub(super) monitor: Vec<MonitorLayoutConfiguration>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub(super) struct MonitorLayoutConfiguration {
+  pub(super) id: String,
+  pub(super) mode: Layout,
+}
+
+/// Window arrangement mode.
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
+#[serde(rename_all = "lowercase")]
+pub enum Layout {
+  /// Free spatial arrangement.
+  #[default]
+  Spatial,
+  /// Workspace-aware horizontal strip.
+  Scrolling,
+}
+
+fn repair_obsolete_empty_monitor_list(config_str: &str) -> Option<String> {
+  if !config_str.contains("[[layout.monitor]]") || !config_str.lines().any(|line| line.trim() == "monitor = []") {
+    return None;
+  }
+
+  Some(
+    config_str
+      .split_inclusive('\n')
+      .filter(|line| line.trim() != "monitor = []")
+      .collect(),
+  )
+}
+
+fn validate_layout_sections(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  let missing_layout = !config_str.contains("[layout]");
+  let missing_spatial = !config_str.contains("[spatial_layout]");
+  let missing_scrolling = !config_str.contains("[scrolling_layout]")
+    || !config_str.contains(SCROLLING_ANIMATION_DURATION_IN_MS)
+    || !config_str.contains(SCROLLING_RECONCILIATION_INTERVAL_IN_MS);
+  if missing_layout || missing_spatial || missing_scrolling {
+    warn!("Layout configuration was incomplete; saving missing defaults");
+    configuration_provider.save_config_or_log_error();
+  }
+
+  if configuration_provider.config.scrolling_layout.animation_duration_in_ms < 0 {
+    warn!(
+      "[{SCROLLING_ANIMATION_DURATION_IN_MS}] is negative, setting to default value: {DEFAULT_SCROLLING_ANIMATION_DURATION_IN_MS}"
+    );
+    configuration_provider.set_i32(SCROLLING_ANIMATION_DURATION_IN_MS, DEFAULT_SCROLLING_ANIMATION_DURATION_IN_MS);
+  }
+  if configuration_provider.config.scrolling_layout.reconciliation_interval_in_ms < 0 {
+    warn!(
+      "[{SCROLLING_RECONCILIATION_INTERVAL_IN_MS}] is negative, setting to default value: {DEFAULT_SCROLLING_RECONCILIATION_INTERVAL_IN_MS}"
+    );
+    configuration_provider.set_i32(
+      SCROLLING_RECONCILIATION_INTERVAL_IN_MS,
+      DEFAULT_SCROLLING_RECONCILIATION_INTERVAL_IN_MS,
+    );
+  }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub(super) struct SpatialLayoutConfiguration {
+  #[serde(default = "default_allow_selecting_same_center_windows")]
+  pub(super) allow_selecting_same_center_windows: bool,
+}
+
+pub(super) fn default_allow_selecting_same_center_windows() -> bool {
+  true
+}
+
+impl Default for SpatialLayoutConfiguration {
+  fn default() -> Self {
+    Self {
+      allow_selecting_same_center_windows: default_allow_selecting_same_center_windows(),
+    }
+  }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub(super) struct ScrollingLayoutConfiguration {
+  #[serde(default = "default_scrolling_animation_duration_in_ms")]
+  pub(super) animation_duration_in_ms: i32,
+  #[serde(default = "default_scrolling_reconciliation_interval_in_ms")]
+  pub(super) reconciliation_interval_in_ms: i32,
+}
+
+fn default_scrolling_animation_duration_in_ms() -> i32 {
+  DEFAULT_SCROLLING_ANIMATION_DURATION_IN_MS
+}
+
+fn default_scrolling_reconciliation_interval_in_ms() -> i32 {
+  DEFAULT_SCROLLING_RECONCILIATION_INTERVAL_IN_MS
+}
+
+impl Default for ScrollingLayoutConfiguration {
+  fn default() -> Self {
+    Self {
+      animation_duration_in_ms: default_scrolling_animation_duration_in_ms(),
+      reconciliation_interval_in_ms: default_scrolling_reconciliation_interval_in_ms(),
+    }
+  }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CustomHotkey {
+  pub name: String,
+  pub path: String,
+  pub hotkey: String,
+  pub execute_as_admin: bool,
+}
+
+/// Settings for excluding certain windows from being managed by the application. This is useful for ignoring
+/// system windows or other applications that should not be affected by this application at all i.e. they should not
+/// be moved, selected, etc.
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ExclusionSettings {
+  #[serde(default = "default_excluded_window_titles")]
+  pub window_titles: Vec<String>,
+  #[serde(default = "default_excluded_window_classes")]
+  pub window_class_names: Vec<String>,
+}
+
+impl Default for ExclusionSettings {
+  fn default() -> Self {
+    Self {
+      window_titles: default_excluded_window_titles(),
+      window_class_names: default_excluded_window_classes(),
+    }
+  }
+}
+
+pub(super) fn default_excluded_window_titles() -> Vec<String> {
+  vec![
+    "Program Manager".to_string(),
+    "Windows Input Experience".to_string(),
+    "".to_string(),
+    "Windows Shell Experience Host".to_string(),
+    "ZPToolBarParentWnd".to_string(),
+    "Settings".to_string(),
+    "ShareX - Screen recording".to_string(),
+    "ShareX - Region capture".to_string(),
+  ]
+}
+
+fn validate_excluded_window_titles(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  if !config_str.contains("window_titles") {
+    warn!(
+      "[{}] was missing; saving it now with default value: {:#?}",
+      "window_titles",
+      default_excluded_window_titles()
+    );
+    configuration_provider.save_config_or_log_error();
+  }
+}
+
+pub(super) fn default_excluded_window_classes() -> Vec<String> {
+  vec![
+    "Progman".to_string(),
+    "WorkerW".to_string(),
+    "Shell_TrayWnd".to_string(),
+    "Shell_SecondaryTrayWnd".to_string(),
+    "DV2ControlHost".to_string(),
+    "#32770".to_string(), // File properties window
+  ]
+}
+
+fn validate_excluded_window_classes(config_str: &str, configuration_provider: &mut ConfigurationProvider) {
+  if !config_str.contains("window_class_names") {
+    warn!(
+      "[{}] was missing; saving it now with default value: {:#?}",
+      "window_class_names",
+      default_excluded_window_classes()
+    );
+    configuration_provider.save_config_or_log_error();
+  }
+}
+
+pub struct ConfigurationProvider {
+  pub(super) file_manager: FileManager<Configuration>,
+  pub(super) config: Configuration,
+}
+
+impl ConfigurationProvider {
+  pub fn new() -> Self {
+    let mut file_manager = FileManager::new(CONFIGURATION_FILE_NAME, FileType::Config);
+    file_manager.set_content_prefix(CONFIGURATION_FILE_PREFIX);
+
+    Self::new_with(file_manager)
+  }
+
+  pub(super) fn new_with(file_manager: FileManager<Configuration>) -> Self {
+    let (config, config_string) = file_manager
+      .load_or_create_with_repair(repair_obsolete_empty_monitor_list)
+      .expect("Failed to load configuration");
+    let mut configuration_provider = ConfigurationProvider { file_manager, config };
+    configuration_provider.validate_config(config_string);
+
+    configuration_provider
+  }
+
+  pub fn log_current_config(&self) {
+    info!("{:?}", self.config);
+  }
+
+  // TODO: Consider validating hotkeys
+  pub(super) fn validate_config(&mut self, config_string: Option<String>) {
+    if let Some(config_as_string) = config_string {
+      validate_window_margin(&config_as_string, self);
+      validate_force_using_admin_privileges(&config_as_string, self);
+      validate_workspace_count(&config_as_string, self);
+      validate_features_using_mouse(&config_as_string, self);
+      validate_delay_in_ms_before_dragging_is_allowed(&config_as_string, self);
+      validate_allow_moving_cursor_after_close_or_minimise(&config_as_string, self);
+      validate_layout_sections(&config_as_string, self);
+      validate_excluded_window_titles(&config_as_string, self);
+      validate_excluded_window_classes(&config_as_string, self);
+    } else {
+      warn!("Failed to validate configuration: configuration string not available");
+    }
+  }
+
+  pub fn get_bool(&self, name: &str) -> bool {
+    match name {
+      ALLOW_SELECTING_SAME_CENTER_WINDOWS => self.config.spatial_layout.allow_selecting_same_center_windows,
+      FORCE_USING_ADMIN_PRIVILEGES => self.config.general.force_using_admin_privileges,
+      ENABLE_FEATURES_USING_MOUSE => self.config.general.enable_features_using_mouse,
+      ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE => {
+        self.config.general.allow_moving_cursor_after_open_close_or_minimise
+      }
+      &_ => {
+        warn!("Failed to get configuration because [{name}] is unknown");
+
+        false
+      }
+    }
+  }
+
+  /// Sets bool value and saves the configuration to file.
+  pub fn set_bool(&mut self, name: &str, value: bool) {
+    match name {
+      ALLOW_SELECTING_SAME_CENTER_WINDOWS => self.config.spatial_layout.allow_selecting_same_center_windows = value,
+      FORCE_USING_ADMIN_PRIVILEGES => self.config.general.force_using_admin_privileges = value,
+      ENABLE_FEATURES_USING_MOUSE => self.config.general.enable_features_using_mouse = value,
+      ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE => {
+        self.config.general.allow_moving_cursor_after_open_close_or_minimise = value
+      }
+      &_ => {
+        warn!("Failed to save configuration because [{name}] is unknown");
+      }
+    }
+    self.save_config_or_log_error();
+  }
+
+  /// Returns the fallback layout for monitors without an override.
+  pub fn get_default_layout(&self) -> Layout {
+    self.config.layout.default
+  }
+
+  /// Saves the fallback layout without changing monitor overrides.
+  pub fn set_default_layout(&mut self, layout: Layout) {
+    self.config.layout.default = layout;
+    self.save_config_or_log_error();
+  }
+
+  /// Resolves a monitor's configured layout.
+  pub fn layout_for_monitor(&self, monitor_id: &str, is_primary: bool) -> Layout {
+    self
+      .config
+      .layout
+      .monitor
+      .iter()
+      .find(|override_| override_.id == monitor_id)
+      .or_else(|| {
+        is_primary
+          .then(|| self.config.layout.monitor.iter().find(|override_| override_.id == "primary"))
+          .flatten()
+      })
+      .map_or(self.config.layout.default, |override_| override_.mode)
+  }
+
+  pub fn get_i32(&self, name: &str) -> i32 {
+    match name {
+      WINDOW_MARGIN => self.config.general.window_margin,
+      ADDITIONAL_WORKSPACE_COUNT => self.config.general.additional_workspace_count,
+      DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED => self.config.general.delay_in_ms_before_dragging_is_allowed,
+      SCROLLING_ANIMATION_DURATION_IN_MS => self.config.scrolling_layout.animation_duration_in_ms,
+      SCROLLING_RECONCILIATION_INTERVAL_IN_MS => self.config.scrolling_layout.reconciliation_interval_in_ms,
+      &_ => {
+        warn!("Failed to get configuration because [{name}] is unknown");
+
+        0
+      }
+    }
+  }
+
+  /// Sets i32 value and saves the configuration to file.
+  pub fn set_i32(&mut self, name: &str, value: i32) {
+    match name {
+      WINDOW_MARGIN => self.config.general.window_margin = value,
+      ADDITIONAL_WORKSPACE_COUNT => self.config.general.additional_workspace_count = value,
+      DELAY_IN_MS_BEFORE_DRAGGING_IS_ALLOWED => self.config.general.delay_in_ms_before_dragging_is_allowed = value,
+      SCROLLING_ANIMATION_DURATION_IN_MS => self.config.scrolling_layout.animation_duration_in_ms = value,
+      SCROLLING_RECONCILIATION_INTERVAL_IN_MS => self.config.scrolling_layout.reconciliation_interval_in_ms = value,
+      &_ => {
+        warn!("Failed to save configuration because [{name}] is unknown");
+      }
+    }
+    self.save_config_or_log_error();
+  }
+
+  pub fn get_hotkeys(&self) -> &Vec<CustomHotkey> {
+    &self.config.hotkey
+  }
+
+  pub fn get_exclusion_settings(&self) -> &ExclusionSettings {
+    &self.config.exclusion_settings
+  }
+
+  pub fn reload_configuration(&mut self) {
+    let (config, config_string) = self
+      .file_manager
+      .load_or_create_with_repair(repair_obsolete_empty_monitor_list)
+      .expect("Failed to reload file");
+    self.config = config;
+    self.validate_config(config_string);
+  }
+
+  fn save_config_or_log_error(&mut self) {
+    if let Err(err) = self.file_manager.save(&self.config) {
+      error!("Failed to save configuration: {}", err);
+    }
+  }
+}
