+++ src/configuration/mod.rs
@@ -1,0 +1,5 @@
+mod configuration_provider;
+#[cfg(test)]
+mod tests;
+
+pub use configuration_provider::*;
