+++ src/configuration/tests/mod.rs
@@ -1,0 +1,3 @@
+use crate::configuration::configuration_provider;
+
+mod configuration_provider_tests;
