+++ src/window_manager/tests/test_support.rs
@@ -1,6 +1,6 @@
 use super::window_manager::WindowManager;
 use crate::api::MockWindowsApi;
-use crate::configuration_provider::{ConfigurationProvider, Layout};
+use crate::configuration::{ConfigurationProvider, Layout};
 use crate::utils::create_temp_directory;
 use crate::workspace::WorkspaceManager;
 use std::sync::{Arc, Mutex};
